loader: Move simple terminators to terminator.c
Keeps the complex terminators in loader.c for the time being.
